requirements.txt
aiohttp
anthropic
boto3
chardet
flask
googlesearch-python
lxml
openai
python-dotenv